Core framework alpha: move all core component code to src.
