+++ packages/core/src/__test__/state.test.ts
